% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{AD_Fulldata}
\alias{AD_Fulldata}
\title{Anaerobic digestion study}
\format{
SummarizedExperiment
}
\source{
The raw data were provided by Dr. Olivier Chapleur and published at
the referenced article. Filtering and normalisation described in our package
vignette.
}
\usage{
data(AD_Fulldata)
}
\description{
This study explored the microbial indicators that could improve the efficacy
of anaerobic digestion (AD) bioprocess and prevent its failure. The samples
were treated with two different ranges of phenol concentration (effect of
interest) and processed at five different dates (batch effect). This study
includes a clear and strong batch effect with an approx. balanced
batch x treatment design.
}
\details{
representation of the raw data as a matrix with 75 samples and 567 OTUs.
}
\note{
See rowData() component of this SummarizedExperiment for the meta data 
information of samples. See colData() component of this SummarizedExperiment 
for detail of the taxonomy of each OTU.
}
\examples{
data(AD_Fulldata)
}
\keyword{datasets}
