% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-se.R
\docType{data}
\name{se_sponge_data}
\alias{se_sponge_data}
\title{Sponge \emph{A. aerophoba} study}
\format{
SummarizedExperiment
}
\source{
The raw data were downloaded from the referenced article. Filtering
and normalisation described in our package vignette.
}
\usage{
data(se_sponge_data)
}
\description{
This study investigated the relationship between metabolite concentration
and microbial abundance of specific sponge tissues. The samples were
collected from two types of tissues (Ectosome vs. Choanosome) and processed
on two separate denaturing gradient gels in electrophoresis. This study
includes relative abundance data only and a completely balanced
batch x treatment design.
}
\details{
representation of the relative abundance data as a data frame with
32 samples and 24 OTUs.
}
\note{
See rowData() component of this SummarizedExperiment for the factor of 
sponge tissues for each sample that is the effect of interest in the sponge 
study and factor of electrophoresis gels where the sample processed for
each sample treated as the batch effect.
}
\examples{
data(se_sponge_data)
}
\keyword{datasets}
