% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sponge_data.r, R/data.r
\docType{data}
\name{sponge_data}
\alias{sponge_data}
\title{Sponge \emph{A. aerophoba} study}
\format{
A list containing four data sets:
\describe{
\item{X.tss}{The relative abundance data represented as a data
frame with 32 samples and 24 OTUs.}
\item{X.clr}{The centered log ratio transformed data including 32 samples
and 24 OTUs from the \code{X.tss}.}
\item{Y.trt}{A factor of sponge tissues for each sample that is the effect
of interest in the sponge study.}
\item{Y.bat}{A factor of electrophoresis gels where the sample processed for
each sample treated as the batch effect.}}

SummarizedExperiment
}
\source{
The raw data were downloaded from the referenced article. Filtering
and normalisation described in our package vignette.

The raw data were downloaded from the referenced article. Filtering
and normalisation described in our package vignette.
}
\usage{
data('sponge_data')

data(sponge_data)
}
\value{
None.
}
\description{
This study investigated the relationship between metabolite concentration
and microbial abundance of specific sponge tissues. The samples were
collected from two types of tissues (Ectosome vs. Choanosome) and processed
on two separate denaturing gradient gels in electrophoresis. This study
includes relative abundance data only and a completely balanced
batch x treatment design.

This study investigated the relationship between metabolite concentration
and microbial abundance of specific sponge tissues. The samples were
collected from two types of tissues (Ectosome vs. Choanosome) and processed
on two separate denaturing gradient gels in electrophoresis. This study
includes relative abundance data only and a completely balanced
batch x treatment design.
}
\details{
representation of the relative abundance data as a data frame with
32 samples and 24 OTUs.
}
\note{
See rowData() component of this SummarizedExperiment for the factor of 
sponge tissues for each sample that is the effect of interest in the sponge 
study and factor of electrophoresis gels where the sample processed for
each sample treated as the batch effect.
}
\examples{
data(sponge_data)
}
\references{
\insertRef{sacristan2011exploring}{PLSDAbatch}
}
\keyword{datasets}
